Pre-emptively abort when approaching rate-limit.
The Github API rate-limits requests to 5000 per hour.  If we hit this
limit part-way through migrating an issue, it could end up permanently
incomplete.  To avoid this, we'll monitor the number of remaining
requests and pre-emptively abort if we're about to migrate an issue
without enough requests held in reserve.