Remove unnecessary escaping of percent signs.
It turns out that the errors when submitting strings containing percent
signs are due to a limitation in Github's API that PyGithub isn't
properly working around.

Since the escaping makes it more difficult to identify issues and
comments that have already been migrated (due to the need to re-apply
the escaping) we'll just assume a patched version of PyGithub and remove
the escaping altogether.